// client/web/src/organisms/mypage/stream-analysis/period-vs-period/PeriodCompareHero.interface.ts
export interface DayStreamsInfo{
  streamId : string;
  title : string;
  platform: 'afreeca'|'youtube'|'twitch';
  airTime: number;
  startedAt: Date;
}

export interface RangeSelectCaledarProps {
  period: Date[];
  handleperiod: (startAt: Date, endAt: Date, base?: true) => void;
  base?: true;
}
export interface periodCompareTextBoxProps {
  base?: true;
  period: Date[];
  handleperiod: (startAt: Date, endAt: Date, base?: true) => void;
}

export interface CheckBoxGroupProps {
  handleCheckStateChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
  viewer: boolean;
  chat: boolean;
  smile: boolean;
}
export interface ISODateTextFieldError {
  helperText: string;
  isError: boolean;
}